+++ frontend/components/App.tsx
@@ -1,4 +1,3 @@
-import { StaticImageData } from "next/image";
 import {
   ChangeEvent,
   createContext,
@@ -10,29 +9,32 @@ import { Fiber, Line } from "../types";
 import Editor from "./Editor";
 import SidePanel from "./SidePanel";
 import SystemMenu from "./SystemMenu";
-import { randomColor } from 'randomcolor';
+import { Props as PendingInferenceProps } from "./InferencePointer";
+import { toast } from "react-toastify";
 
 interface State {
   isValidScale: boolean;
   magnitude: string;
   scaleLength: number;
   realDims: { width: number; height: number };
-  isChoosingTarget: boolean;
-  imagePath: string;
-}
-
-interface AppContextType {
-  imageDimsInPx: {
+  htmlImageDims: {
     width: number;
     height: number;
     x: number;
     y: number;
   };
+  isChoosingTarget: boolean;
+  imagePath: string;
+  pendingInferences: PendingInferenceProps[];
+}
+
+interface AppContextType {
   fibers: Fiber[];
   setFibers: Dispatch<SetStateAction<Fiber[]>>;
+  addFiber: (measurements: Line[], color: string) => void;
+  swapImage: (event: ChangeEvent<HTMLInputElement>) => void;
   appState: State;
   setAppState: Dispatch<SetStateAction<State>>;
-  addFiber: (measurements: Line[]) => void;
 }
 
 export const AppContext = createContext<AppContextType | undefined>(undefined);
@@ -47,8 +49,10 @@ const App = () => {
     magnitude: "nm",
     scaleLength: 400,
     realDims: { width: 0, height: 0 },
+    htmlImageDims: { width: 0, height: 0, x: 0, y: 0 },
     isChoosingTarget: false,
-    imagePath: "/images/fibers.png",
+    imagePath: '',
+    pendingInferences: [] as PendingInferenceProps[],
   });
 
   const onScaleChange = (event: ChangeEvent<HTMLInputElement>) => {
@@ -61,30 +65,49 @@ const App = () => {
     }
   };
 
-  const addFiber = (measurements: Line[]) => {
-     setFibers((prevFibers) => {
-       const id = Math.max(0, ...fibers.map((fiber) => fiber.id)) + 1;
-       prevFibers.push({
-         id,
-         color: randomColor(),
-         measurements,
-       });
-       return [...prevFibers];
-     });
-  }
+  const addFiber = (measurements: Line[], color: string) => {
+    console.log("add fiber")
+    setFibers((prevFibers) => {
+      const id = Math.max(0, ...prevFibers.map((fiber) => fiber.id)) + 1;
+      console.log("push fiber")
+      prevFibers.push({
+        id,
+        color,
+        measurements,
+      });
+      return [...prevFibers];
+    });
+  };
+
+  const swapImage = (event: ChangeEvent<HTMLInputElement>) => {
+    if (!(event.target.files && event.target.files[0])) {
+      toast.error("no file selected");
+      return;
+    }
+
+    const file = event.target.files[0];
+    if (file.type !== "image/png" && file.type !== "image/jpeg") {
+      toast.error("This only accepts PNG or JPG images");
+      return;
+    }
+
+    const imagePath = URL.createObjectURL(file);
+    
+    setFibers([]);
+    setState((prevState) => ({
+      ...prevState,
+      imagePath,
+    }));
+  };
+  
   return (
-    <div className='container'>
+    <div className='h-screen py-6 bg-black'>
       <AppContext.Provider
         value={{
-          imageDimsInPx: {
-            width: 0,
-            height: 0,
-            x: 0,
-            y: 0,
-          },
           fibers,
           setFibers,
           addFiber,
+          swapImage,
           appState: state,
           setAppState: setState,
         }}